# Better hash