Towards fixing https://github.com/harfbuzz/harfbuzz/issues/4227