@behdad 2023-05-08
Number of probes by length:

Before:
```
      4 61
      4 62
      8 60
     20 59
     24 58
     34 57
     46 56
     62 55
     64 54
     78 53
     92 52
    102 51
    116 50
    138 49
    156 48
    190 47
    218 46
    252 45
    294 44
    344 43
    386 42
    444 41
    494 40
    586 39
    704 38
    758 37
    788 36
    868 35
    976 34
   1060 33
   1137 32
   1200 31
   1297 30
   1394 29
   1479 28
   1533 27
   1586 26
   1639 25
   1693 24
   1761 23
   1836 22
   1918 21
   2044 20
   2138 19
   2245 18
   2356 17
   2483 16
   2614 15
   2750 14
   3295 13
   3688 12
   4578 11
   5065 10
   6435 9
   7220 8
   9518 7
  11356 6
  14195 5
  18165 4
  22666 3
  31777 2
1125146 1
```

After:
```
      2 19
      2 20
      2 21
      2 22
      2 23
      4 18
      6 17
     23 16
     32 15
     44 14
     54 13
     83 12
    144 11
    226 10
    357 9
    867 8
   1219 7
   2152 6
   4022 5
   8224 4
  19188 3
  55419 2
1119223 1
```

Other than the length reduction, the number of long probes drops down rather quickly.