@behdad 2023-05-08
While these improve the worst-case performance of the hashmap, I do see 6% slowdown on this benchmark case though. So I'm undecided. :(