@behdad 2023-05-09
With fasthash. Not that different. Perhaps because saturated by int hash.
```
      2 19
      2 20
      6 17
      6 18
     12 16
     17 15
     31 14
     51 13
     90 12
    134 11
    206 10
    329 9
    614 8
    947 7
   1690 6
   3008 5
   7432 4
  18397 3
  53821 2
1119453 1
```