@behdad 2023-05-09
With the latest change the performance seems to be a net gain slightly. New numbers:
```
      2 19
      2 20
      8 18
     22 17
     28 16
     34 15
     50 14
     73 13
    127 12
    183 11
    272 10
    419 9
    755 8
   1132 7
   1731 6
   3131 5
   5574 4
   9357 3
  36606 2
1116277 1
```