@behdad 2023-05-09
Humm. Just for `hb_bytes_t` I get the following:
```
      2 19
      2 20
      6 17
      6 18
     12 16
     16 15
     30 14
     50 13
     89 12
    133 11
    198 10
    315 9
    588 8
    895 7
   1413 6
   2592 5
   4748 4
   7981 3
  14511 2
  37638 1
```