@behdad 2023-05-09
> Humm. Just for `hb_bytes_t` I get the following:

This is still surprisingly long, given the small number of operations.