@behdad 2023-05-10
I ended up disabling the integer hash function since it has negative performance effect.