@behdad 2023-05-26
> I ended up disabling the integer hash function since it has negative performance effect.

1. The subset benchmarks show a slowdown of ~5% according to @garretrieger 

2. The `benchmark-map` diff is a lot more interesting:
```
Comparing before to after
Benchmark                              Time             CPU      Time Old      Time New       CPU Old       CPU New
-------------------------------------------------------------------------------------------------------------------
BM_MapInsert/16                     +0.9944         +0.9949            27            53            26            53
BM_MapInsert/64                     +0.5324         +0.5328            34            53            34            53
BM_MapInsert/512                    +0.2256         +0.2259            43            53            43            53
BM_MapInsert/4096                   -0.5521         -0.5520           118            53           118            53
BM_MapInsert/32768                  -0.7956         -0.7956           259            53           258            53
BM_MapInsert/262144                 -0.8242         -0.8243           316            55           315            55
BM_MapInsert/1048576                -0.8764         -0.8764           647            80           645            80
BM_MapLookup/16                     +0.0975         +0.0975             3             3             3             3
BM_MapLookup/64                     +0.1666         +0.1663             3             4             3             4
BM_MapLookup/512                    +0.2666         +0.2686             3             4             3             4
BM_MapLookup/4096                   +1.7377         +1.7374             3             9             3             9
BM_MapLookup/32768                  +0.3670         +0.3671             7             9             7             9
BM_MapLookup/262144                 +0.9840         +0.9854             7            14             7            14
BM_MapLookup/1048576                +1.1744         +1.1729             7            15             7            15
OVERALL_GEOMEAN                     -0.1102         -0.1101             0             0             0             0
```
Let's look at the before/after separately:

Before:
```
---------------------------------------------------------------
Benchmark                     Time             CPU   Iterations
---------------------------------------------------------------
BM_MapInsert/16            26.6 ns         26.5 ns     33423282
BM_MapInsert/64            34.4 ns         34.3 ns     28868762
BM_MapInsert/512           43.0 ns         42.8 ns     20134729
BM_MapInsert/4096           118 ns          118 ns      5452951
BM_MapInsert/32768          259 ns          258 ns      4198637
BM_MapInsert/262144         316 ns          315 ns      1773352
BM_MapInsert/1048576        647 ns          645 ns      1408436
BM_MapLookup/16            3.10 ns         3.09 ns    226235463
BM_MapLookup/64            3.19 ns         3.18 ns    220335005
BM_MapLookup/512           3.33 ns         3.32 ns    206349995
BM_MapLookup/4096          3.47 ns         3.46 ns    198738054
BM_MapLookup/32768         6.86 ns         6.84 ns    103205802
BM_MapLookup/262144        6.90 ns         6.87 ns    102260716
BM_MapLookup/1048576       6.85 ns         6.84 ns    102540738
```
After:
```
---------------------------------------------------------------
Benchmark                     Time             CPU   Iterations
---------------------------------------------------------------
BM_MapInsert/16            53.0 ns         52.8 ns     10000000
BM_MapInsert/64            52.7 ns         52.5 ns     10000000
BM_MapInsert/512           52.7 ns         52.5 ns     10000000
BM_MapInsert/4096          52.8 ns         52.7 ns     10000000
BM_MapInsert/32768         53.0 ns         52.8 ns     10000000
BM_MapInsert/262144        55.5 ns         55.3 ns     10000000
BM_MapInsert/1048576       80.0 ns         79.8 ns     11023447
BM_MapLookup/16            3.40 ns         3.39 ns    206079247
BM_MapLookup/64            3.72 ns         3.71 ns    189178653
BM_MapLookup/512           4.22 ns         4.21 ns    166123694
BM_MapLookup/4096          9.50 ns         9.48 ns     73851363
BM_MapLookup/32768         9.38 ns         9.36 ns     74429576
BM_MapLookup/262144        13.7 ns         13.6 ns     51010831
BM_MapLookup/1048576       14.9 ns         14.9 ns     47329274
```

We observe two things, one good and one bad:

== The good

The MapInsert performance used to get worse as the initial number of items in the map increased:
```
BM_MapInsert/16            26.6 ns         26.5 ns     33423282
BM_MapInsert/64            34.4 ns         34.3 ns     28868762
BM_MapInsert/512           43.0 ns         42.8 ns     20134729
BM_MapInsert/4096           118 ns          118 ns      5452951
BM_MapInsert/32768          259 ns          258 ns      4198637
BM_MapInsert/262144         316 ns          315 ns      1773352
BM_MapInsert/1048576        647 ns          645 ns      1408436
```
That's because of hitting long chains (square-root?) with the poor integer hash function. With the improved (Knuth's multiplicative) hash function, the time is even across the benchmark:
```
BM_MapInsert/16            53.0 ns         52.8 ns     10000000
BM_MapInsert/64            52.7 ns         52.5 ns     10000000
BM_MapInsert/512           52.7 ns         52.5 ns     10000000
BM_MapInsert/4096          52.8 ns         52.7 ns     10000000
BM_MapInsert/32768         53.0 ns         52.8 ns     10000000
BM_MapInsert/262144        55.5 ns         55.3 ns     10000000
BM_MapInsert/1048576       80.0 ns         79.8 ns     11023447
```

== The bad

The MapLookup performance used to be this:
```
BM_MapLookup/16            3.10 ns         3.09 ns    226235463
BM_MapLookup/64            3.19 ns         3.18 ns    220335005
BM_MapLookup/512           3.33 ns         3.32 ns    206349995
BM_MapLookup/4096          3.47 ns         3.46 ns    198738054
BM_MapLookup/32768         6.86 ns         6.84 ns    103205802
BM_MapLookup/262144        6.90 ns         6.87 ns    102260716
BM_MapLookup/1048576       6.85 ns         6.84 ns    102540738
```

Now it's this:
```
BM_MapLookup/16            3.40 ns         3.39 ns    206079247
BM_MapLookup/64            3.72 ns         3.71 ns    189178653
BM_MapLookup/512           4.22 ns         4.21 ns    166123694
BM_MapLookup/4096          9.50 ns         9.48 ns     73851363
BM_MapLookup/32768         9.38 ns         9.36 ns     74429576
BM_MapLookup/262144        13.7 ns         13.6 ns     51010831
BM_MapLookup/1048576       14.9 ns         14.9 ns     47329274
```

So, instead of, roughly, 3,3,3,3,7,7,7; we now get 4,4,4,9,9,14,14.  After lots of studying (with `perf stat` and `cachegrind`) I'm convinced that this is just memory cache latency of the more random read access with the better hash function. The test case does map lookups of successive integers, which is very advantageous to the identity hash function.

So. Mystery solved. I think we might want to use the better hash function after all, to avoid long chains.